change flatten function location to crossfire.clients.occurrences
