+++ crossfire/clients/occurrences.py
@@ -6,10 +6,15 @@ from urllib.parse import urlencode
 from httpx import ReadTimeout
 from tqdm import tqdm
 
+from crossfire.errors import NestedColumnError
+
 try:
-    from pandas import concat
+    from pandas import DataFrame, Series, concat
+
+    HAS_PANDAS = True
 except ImportError:
-    pass
+    HAS_PANDAS = False
+
 
 try:
     from geopandas import GeoDataFrame
@@ -28,6 +33,14 @@ logger = Logger(__name__)
 
 TYPE_OCCURRENCES = {"all", "withVictim", "withoutVictim"}
 NOT_NUMBER = re.compile("\D")
+NESTED_COLUMNS = {
+    "contextInfo",
+    "state",
+    "region",
+    "city",
+    "neighborhood",
+    "locality",
+}
 
 
 def date_formatter(date_parameter):
@@ -164,3 +177,43 @@ class Accumulator:
             return GeoDataFrame(self.data)
 
         return self.data
+
+
+def _flatten_df(row, column_name):
+    column_data = row[column_name]
+    return Series(
+        {f"{column_name}_{key}": value for key, value in column_data.items()}
+    )
+
+
+def is_empty(data):
+    if HAS_PANDAS and isinstance(data, DataFrame):
+        return data.empty
+    return not data
+
+
+def flatten(data, nested_columns=None):
+    nested_columns = set(nested_columns or NESTED_COLUMNS)
+    if not nested_columns.issubset(NESTED_COLUMNS):
+        raise NestedColumnError(nested_columns)
+    if is_empty(data):
+        return data
+    if HAS_PANDAS and isinstance(data, DataFrame):
+        keys = set(data.columns) & nested_columns
+        for key in keys:
+            data = concat(
+                (
+                    data.drop(key, axis=1),
+                    data.apply(_flatten_df, args=(key,), axis=1),
+                ),
+                axis=1,
+            )
+
+        return data
+
+    keys = set(data[0].keys()) & nested_columns
+    for item in data:
+        for key in keys:
+            item.update({f"{key}_{k}": v for k, v in item.get(key).items()})
+            item.pop(key)
+    return data
